pyproject.toml
[project]
name = "demo-sistem"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.1.0",
    "langchain-classic>=1.0.0",
    "langchain-community>=0.3.31",
    "langchain-core>=0.3.79",
    "langchain-huggingface>=1.1.0",
    "langchain-openai>=0.3.35",
    "langchain-postgres>=0.0.16",
    "openai>=2.8.1",
    "pandas>=2.3.3",
    "psycopg>=3.3.0",
    "psycopg2>=2.9.11",
    "pypdf>=6.4.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "sqlalchemy>=2.0.44",
    "streamlit>=1.51.0",
]
